backend/requirements.txt
# ============================================================
# JustorAI RAG Brain — Backend Dependencies
# Python 3.14 / Windows compatible
# ============================================================

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.8.0
python-multipart>=0.0.9

# Env vars
python-dotenv>=1.0.0

# HTTP (used by ingest.py for uploading PDFs)
requests>=2.31.0

# Supabase — pin sub-packages to avoid storage3's pyiceberg dep
# storage3 added pyiceberg in 2.10.0; pin below that
supabase>=2.7.4,<2.10.0
httpx>=0.27.0

# Embeddings — all-MiniLM-L6-v2 (384-dim, ~80 MB, CPU only)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers>=2.7.0

# PDF parsing
PyPDF2>=3.0.1

# Groq LLM
groq>=0.9.0
websockets>=13.0